modest to upraised
It's true that her terrace is modest, but really what it is is raised up above the street in front of it. The whole city is built on kind of a large hill, so the buildings are tiered, one row above the next, and they're built up high to prevent unwanted company from entering easily.